cleanUpConfig
