+++ src/css/config.css
@@ -76,7 +76,7 @@
 
 .clone-repo a {
   text-decoration: none;
-  color: #fff;
+  color: #00a4ef;
   font-weight: 400;
   opacity: 0.7;
   font-size: 0.8rem;
